# swagger_server/models/payment_method.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class PaymentMethod(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, card_number: str=None, expire_month: int=None, expire_year: int=None, card_holder: str=None):  # noqa: E501
        """PaymentMethod - a model defined in Swagger

        :param card_number: The card_number of this PaymentMethod.  # noqa: E501
        :type card_number: str
        :param expire_month: The expire_month of this PaymentMethod.  # noqa: E501
        :type expire_month: int
        :param expire_year: The expire_year of this PaymentMethod.  # noqa: E501
        :type expire_year: int
        :param card_holder: The card_holder of this PaymentMethod.  # noqa: E501
        :type card_holder: str
        """
        self.swagger_types = {
            'card_number': str,
            'expire_month': int,
            'expire_year': int,
            'card_holder': str
        }

        self.attribute_map = {
            'card_number': 'cardNumber',
            'expire_month': 'expireMonth',
            'expire_year': 'expireYear',
            'card_holder': 'cardHolder'
        }
        self._card_number = card_number
        self._expire_month = expire_month
        self._expire_year = expire_year
        self._card_holder = card_holder

    @classmethod
    def from_dict(cls, dikt) -> 'PaymentMethod':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The PaymentMethod of this PaymentMethod.  # noqa: E501
        :rtype: PaymentMethod
        """
        return util.deserialize_model(dikt, cls)

    @property
    def card_number(self) -> str:
        """Gets the card_number of this PaymentMethod.

        Masked card number.  # noqa: E501

        :return: The card_number of this PaymentMethod.
        :rtype: str
        """
        return self._card_number

    @card_number.setter
    def card_number(self, card_number: str):
        """Sets the card_number of this PaymentMethod.

        Masked card number.  # noqa: E501

        :param card_number: The card_number of this PaymentMethod.
        :type card_number: str
        """
        if card_number is None:
            raise ValueError("Invalid value for `card_number`, must not be `None`")  # noqa: E501

        self._card_number = card_number

    @property
    def expire_month(self) -> int:
        """Gets the expire_month of this PaymentMethod.

        Expire month of the card.  # noqa: E501

        :return: The expire_month of this PaymentMethod.
        :rtype: int
        """
        return self._expire_month

    @expire_month.setter
    def expire_month(self, expire_month: int):
        """Sets the expire_month of this PaymentMethod.

        Expire month of the card.  # noqa: E501

        :param expire_month: The expire_month of this PaymentMethod.
        :type expire_month: int
        """
        if expire_month is None:
            raise ValueError("Invalid value for `expire_month`, must not be `None`")  # noqa: E501

        self._expire_month = expire_month

    @property
    def expire_year(self) -> int:
        """Gets the expire_year of this PaymentMethod.

        Expire year of the card.  # noqa: E501

        :return: The expire_year of this PaymentMethod.
        :rtype: int
        """
        return self._expire_year

    @expire_year.setter
    def expire_year(self, expire_year: int):
        """Sets the expire_year of this PaymentMethod.

        Expire year of the card.  # noqa: E501

        :param expire_year: The expire_year of this PaymentMethod.
        :type expire_year: int
        """
        if expire_year is None:
            raise ValueError("Invalid value for `expire_year`, must not be `None`")  # noqa: E501

        self._expire_year = expire_year

    @property
    def card_holder(self) -> str:
        """Gets the card_holder of this PaymentMethod.

        Name of the card  # noqa: E501

        :return: The card_holder of this PaymentMethod.
        :rtype: str
        """
        return self._card_holder

    @card_holder.setter
    def card_holder(self, card_holder: str):
        """Sets the card_holder of this PaymentMethod.

        Name of the card  # noqa: E501

        :param card_holder: The card_holder of this PaymentMethod.
        :type card_holder: str
        """
        if card_holder is None:
            raise ValueError("Invalid value for `card_holder`, must not be `None`")  # noqa: E501

        self._card_holder = card_holder