# swagger_server/models/purchase.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Purchase(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, purchase_price: float=None, purchase_date: datetime=None, payment_method_id: int=None, song_ids: List[int]=None, album_ids: List[int]=None, merch_ids: List[int]=None):  # noqa: E501
        """Purchase - a model defined in Swagger

        :param purchase_price: The purchase_price of this Purchase.  # noqa: E501
        :type purchase_price: float
        :param purchase_date: The purchase_date of this Purchase.  # noqa: E501
        :type purchase_date: datetime
        :param payment_method_id: The payment_method_id of this Purchase.  # noqa: E501
        :type payment_method_id: int
        :param song_ids: The song_ids of this Purchase.  # noqa: E501
        :type song_ids: List[int]
        :param album_ids: The album_ids of this Purchase.  # noqa: E501
        :type album_ids: List[int]
        :param merch_ids: The merch_ids of this Purchase.  # noqa: E501
        :type merch_ids: List[int]
        """
        self.swagger_types = {
            'purchase_price': float,
            'purchase_date': datetime,
            'payment_method_id': int,
            'song_ids': List[int],
            'album_ids': List[int],
            'merch_ids': List[int]
        }

        self.attribute_map = {
            'purchase_price': 'purchasePrice',
            'purchase_date': 'purchaseDate',
            'payment_method_id': 'paymentMethodId',
            'song_ids': 'songIds',
            'album_ids': 'albumIds',
            'merch_ids': 'merchIds'
        }
        self._purchase_price = purchase_price
        self._purchase_date = purchase_date
        self._payment_method_id = payment_method_id
        self._song_ids = song_ids
        self._album_ids = album_ids
        self._merch_ids = merch_ids

    @classmethod
    def from_dict(cls, dikt) -> 'Purchase':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Purchase of this Purchase.  # noqa: E501
        :rtype: Purchase
        """
        return util.deserialize_model(dikt, cls)

    @property
    def purchase_price(self) -> float:
        """Gets the purchase_price of this Purchase.

        Price of the Song  # noqa: E501

        :return: The purchase_price of this Purchase.
        :rtype: float
        """
        return self._purchase_price

    @purchase_price.setter
    def purchase_price(self, purchase_price: float):
        """Sets the purchase_price of this Purchase.

        Price of the Song  # noqa: E501

        :param purchase_price: The purchase_price of this Purchase.
        :type purchase_price: float
        """
        if purchase_price is None:
            raise ValueError("Invalid value for `purchase_price`, must not be `None`")  # noqa: E501

        self._purchase_price = purchase_price

    @property
    def purchase_date(self) -> datetime:
        """Gets the purchase_date of this Purchase.

        Date of the purchase  # noqa: E501

        :return: The purchase_date of this Purchase.
        :rtype: datetime
        """
        return self._purchase_date

    @purchase_date.setter
    def purchase_date(self, purchase_date: datetime):
        """Sets the purchase_date of this Purchase.

        Date of the purchase  # noqa: E501

        :param purchase_date: The purchase_date of this Purchase.
        :type purchase_date: datetime
        """
        if purchase_date is None:
            raise ValueError("Invalid value for `purchase_date`, must not be `None`")  # noqa: E501

        self._purchase_date = purchase_date

    @property
    def payment_method_id(self) -> int:
        """Gets the payment_method_id of this Purchase.

        id of the payment method  # noqa: E501

        :return: The payment_method_id of this Purchase.
        :rtype: int
        """
        return self._payment_method_id

    @payment_method_id.setter
    def payment_method_id(self, payment_method_id: int):
        """Sets the payment_method_id of this Purchase.

        id of the payment method  # noqa: E501

        :param payment_method_id: The payment_method_id of this Purchase.
        :type payment_method_id: int
        """
        if payment_method_id is None:
            raise ValueError("Invalid value for `payment_method_id`, must not be `None`")  # noqa: E501

        self._payment_method_id = payment_method_id

    @property
    def song_ids(self) -> List[int]:
        """Gets the song_ids of this Purchase.

        List of song IDs to purchase  # noqa: E501

        :return: The song_ids of this Purchase.
        :rtype: List[int]
        """
        return self._song_ids

    @song_ids.setter
    def song_ids(self, song_ids: List[int]):
        """Sets the song_ids of this Purchase.

        List of song IDs to purchase  # noqa: E501

        :param song_ids: The song_ids of this Purchase.
        :type song_ids: List[int]
        """
        self._song_ids = song_ids

    @property
    def album_ids(self) -> List[int]:
        """Gets the album_ids of this Purchase.

        List of album IDs to purchase  # noqa: E501

        :return: The album_ids of this Purchase.
        :rtype: List[int]
        """
        return self._album_ids

    @album_ids.setter
    def album_ids(self, album_ids: List[int]):
        """Sets the album_ids of this Purchase.

        List of album IDs to purchase  # noqa: E501

        :param album_ids: The album_ids of this Purchase.
        :type album_ids: List[int]
        """
        self._album_ids = album_ids

    @property
    def merch_ids(self) -> List[int]:
        """Gets the merch_ids of this Purchase.

        List of merch IDs to purchase  # noqa: E501

        :return: The merch_ids of this Purchase.
        :rtype: List[int]
        """
        return self._merch_ids

    @merch_ids.setter
    def merch_ids(self, merch_ids: List[int]):
        """Sets the merch_ids of this Purchase.

        List of merch IDs to purchase  # noqa: E501

        :param merch_ids: The merch_ids of this Purchase.
        :type merch_ids: List[int]
        """
        self._merch_ids = merch_ids